requirements.txt
streamlit
tweepy
Pillow
python-dotenv
requests
